// Web/LearningStarter/learning-starter-web/src/constants/types.tsx
//This type uses a generic (<T>).  For more information on generics see: https://www.typescriptlang.org/docs/handbook/2/generics.html
//You probably wont need this for the scope of this class :)
export type ApiResponse<T> = {
  data: T;
  errors: ApiError[];
  hasErrors: boolean;
};

export type ApiError = {
  property: string;
  message: string;
};

export type AnyObject = {
  [index: string]: any;
};

export type UserDto = {
  id: number;
  firstName: string;
  lastName: string;
  userName: string;
};

export type   ServersGetDto = {
  id: number;
  name: string;
  description: string;
  memberCount: number;
  createdAt: string;
};

export type   landingpageGetDto = {
  
};

export type ServersCreateUpdateDto= {

};